split fg class based on functionality
